reorg unified-doc and unified-doc-parse-text packages
