--- docs/packages/processor.mdx
@@ -1,9 +1,0 @@
----
-name: processor
-menu: Packages
-route: packages/processor
----
-
-import Readme from "@packages/processor/readme.md";
-
-<Readme />
