--- docs/packages/text-parse.mdx
@@ -1,9 +1,0 @@
----
-name: text-parse
-menu: Packages
-route: packages/text-parse
----
-
-import Readme from "@packages/text-parse/readme.md";
-
-<Readme />
