+++ docs/packages/unified-doc-parse-text.mdx
@@ -1,0 +1,8 @@
+---
+name: unified-doc-parse-text
+route: packages/unified-doc-parse-text
+---
+
+import Readme from "@packages/unified-doc-parse-text/readme";
+
+<Readme />
