+++ docs/packages/unified-doc.mdx
@@ -1,0 +1,8 @@
+---
+name: unified-doc
+route: packages/unified-doc
+---
+
+import Readme from "@packages/unified-doc/readme";
+
+<Readme />
