--- packages/text-parse/src/index.d.ts
@@ -1,5 +1,0 @@
-import { Node } from 'unist';
-
-export function parser(doc: string): Node;
-
-export default function parse(): void;
