+++ packages/unified-doc-parse-text/index.js
@@ -1,0 +1,1 @@
+export { default } from './src/parse';
