+++ packages/unified-doc-parse-text/package.json
@@ -1,6 +1,6 @@
 {
-	"name": "@unified-doc/text-parse",
-	"version": "0.1.1",
+	"name": "unified-doc-parse-text",
+	"version": "1.0.0",
 	"description": "unified parser to parse text content into a hast tree with a single text node",
 	"main": "dist/index.js",
 	"module": "dist/index.module.js",
@@ -9,8 +9,9 @@
 	"keywords": [
 		"unified",
 		"doc",
-		"text",
 		"parse",
+		"text",
+		"document",
 		"content",
 		"unist",
 		"hast",
@@ -18,7 +19,7 @@
 	],
 	"author": "Chris Zhou <chrisrzhou@pm.me>",
 	"license": "MIT",
-	"repository": "https://github.com/chrisrzhou/unified-doc/tree/master/packages/react-unified-doc",
+	"repository": "https://github.com/chrisrzhou/unified-doc/tree/master/packages/unified-doc-parse-text",
 	"bugs": "https://github.com/chrisrzhou/unified-doc/issues",
 	"homepage": "https://unified-doc.netlify.com/",
 	"scripts": {
