+++ packages/unified-doc-parse-text/src/parse.d.ts
@@ -1,0 +1,1 @@
+export default function parse(): void;
