+++ packages/unified-doc-parse-text/src/parse.js
@@ -1,0 +1,5 @@
+import parser from './parser';
+
+export default function parse() {
+	this.Parser = parser;
+}
