+++ packages/unified-doc-parse-text/src/parser.d.ts
@@ -1,0 +1,3 @@
+import { Node } from 'unist';
+
+export default function parser(doc: string): Node;
