+++ packages/unified-doc/readme.md
@@ -1,0 +1,3 @@
+# unified-doc
+
+`unified-doc` APIs and typings.
