+++ packages/unified-doc/src/index.d.ts
@@ -1,0 +1,5 @@
+import createProcessor from './create-processor';
+
+export * from './types';
+
+export { createProcessor };
