+++ packages/unified-doc/src/index.js
@@ -1,0 +1,1 @@
+export { default as createProcessor } from './create-processor';
