+++ packages/unified-doc/src/types.d.ts
@@ -1,0 +1,1 @@
+export type ContentType = 'html' | 'markdown' | 'text';
