+++ packages/unified-doc/src/utils/coerce-text-positions.d.ts
